requirements.txt
requests
pyquery
js2py
grequests
pymysql
sqlalchemy
jaconv
tornado
boto3
mtranslate
pillow
vcrpy
